% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyFit-package.R
\docType{package}
\name{anyFit-package}
\alias{anyFit}
\alias{anyFit-package}
\title{anyFit: anyFit: Exploratory Data Analysis on xts timeseries objects and spatial netCDF objects}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Package for exploratory data analysis, timeseries analysis and visualisation of point and spatially distributed environmental variables.
}
\details{
anyFit streamlines exploratory data analysis of point and gridded
hydroclimatic time series: loading (NetCDF and delimited files), summary
statistics, temporal aggregation, and distribution fitting by the method of
L-moments, all built on the \pkg{xts} time-series format and \pkg{ggplot2}
visualisation. Gridded data are carried in the \link{sxts} (spatial xts) class,
which stores coordinates and projection alongside the series.

The package ships a real example dataset used throughout the vignettes and
examples: an E-OBS ensemble-mean daily rainfall grid over Europe
(0.25 degrees, 2011-2022), installed at
\code{system.file("extdata", "rr_ens_mean_0.25deg_reg_2011-2022_v27.0e.nc", package = "anyFit")}.
Load it with \code{\link[=nc2xts]{nc2xts()}}, typically clipped to a region, e.g.
\code{nc2xts(f, "rr", country = "U.K. of Great Britain and Northern Ireland")}.
A daily station series in
\code{system.file("extdata", "KNMI_Daily.csv", package = "anyFit")}
illustrates \code{\link[=delim2xts]{delim2xts()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gapouliasis.github.io/anyFit/}
  \item \url{https://github.com/Gapouliasis/anyFit}
  \item Report bugs at \url{https://github.com/Gapouliasis/anyFit/issues}
}

}
\author{
\strong{Maintainer}: George Pouliasis \email{gapouliasis@gmail.com}

Authors:
\itemize{
  \item George Pouliasis \email{gapouliasis@gmail.com}
}

}
\keyword{internal}
