% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{GenPareto}
\alias{GenPareto}
\alias{dgpd}
\alias{pgpd}
\alias{qgpd}
\alias{rgpd}
\title{Generalized Pareto Distribution}
\usage{
dgpd(x, location, scale, shape)

pgpd(q, location, scale, shape)

qgpd(p, location, scale, shape)

rgpd(n, location, scale, shape)
}
\description{
Generalized Pareto distribution
}
