% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxts.R
\name{mask.sxts}
\alias{mask.sxts}
\title{Spatial masking of sxts objects}
\usage{
mask.sxts(obj, xlim = NA, ylim = NA, shapefile_name = NA, mask = NA)
}
\description{
Spatially subsets an sxts object by bounding box (\code{xlim},
\code{ylim}), country name, continent name, or a shapefile. When using
polygons, point-in-polygon membership is determined via
\code{sf::st_within} with CRS-aware coordinate transformation between the
sxts CRS and the mask CRS. Points falling outside the mask are dropped.
}
