% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxts.R
\name{rasterFromSxts}
\alias{rasterFromSxts}
\title{Convert sxts to RasterBrick}
\usage{
rasterFromSxts(x, ...)
}
\arguments{
\item{x}{An sxts object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
S3 generic for converting an sxts object to a \code{RasterBrick}.
}
