% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_xts.R
\name{aggregate_xts}
\alias{aggregate_xts}
\title{aggregate_xts}
\usage{
aggregate_xts(
  ts,
  periods = NA,
  FUN = "mean",
  period_multiplier = NA,
  pstart = NA,
  pend = NA,
  season_end = NA,
  mseason_title = NA
)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{periods}{A vector of periods to aggregate the timeseries to.
Accepts 'mins' or 'minutes', 'hours', 'days', 'weeks', 'months', 'quarters', and 'years'.}

\item{FUN}{The function used to aggregate the data. Defaults to 'mean'.}

\item{period_multiplier}{Multiplier to aggregate data to a non-standard time scale.
Must be a vector with the same length as periods.}

\item{pstart}{Plotting start date. If not included, the first date of the timeseries will be used.
Usefull for long timeseries.}

\item{pend}{Plotting end date. If not included, the last date of the timeseries will be used.
Usefull for long timeseries.}

\item{season_end}{Vector of season end dates. Can be used to aggregate into user-defined seasons.
Must be in the form of '\%m-\%d', e.g. '10-15'}

\item{mseason_title}{Title for the user-defined season.}
}
\value{
A list with the aggregated timeseries for every selected period in xts format and
a plot containing the aggregated timeseries.
}
\description{
Aggregates a xts timeseries to different time scales
and returns a list of plots and aggregated timeseries.
}
\examples{
file <- "KNMI_Daily.csv"
file_path <- file.path(cwd,file)
time_zone <- "UTC"
time_step <- "1 day"

data <- delim2xts(file_path = file_path, 
                 time_zone = "UTC", delim = " ", time_step = time_step)
                
agg_ts <- aggregate_xts(data[,4], c("months","quarters","years"), FUN = 'sum')
smonthly_ts <- agg_ts$list_months$aggregated                

}
