% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{basic_stats}
\alias{basic_stats}
\title{basic_stats}
\usage{
basic_stats(
  ts,
  pstart = NA,
  pend = NA,
  show_label = TRUE,
  label_prefix = "Station",
  show_table = TRUE,
  xpos_label = 0.1,
  ypos_label = 0.95,
  xpos_table = 0.1,
  ypos_table = 0.15,
  nbins = 30,
  ignore_zeros = FALSE,
  zero_threshold = 0.01,
  plot = FALSE
)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{pstart}{Plotting start date. If not included, the first date of the timeseries will be used.
Usefull for long timeseries.}

\item{pend}{Plotting end date. If not included, the last date of the timeseries will be used.
Usefull for long timeseries.}

\item{show_label}{A logical value, if TRUE the timeseries title will be plotted.
It is assumed to be the column name. Default is TRUE.}

\item{label_prefix}{A character value wich contains the prefix of the timeseries title. Default is 'station'}

\item{show_table}{A logical value, if TRUE a table with the mean value, the standard deviation, the skewness and
the probability dry is plotted together with the timeseries. Default is true.}

\item{xpos_label}{The x position of the timeseries label. Takes values from 0 to 1.}

\item{ypos_label}{The y position of the timeseries label. Takes values from 0 to 1.}

\item{xpos_table}{The x position of the timeseries table with the basic statistics. Takes values from 0 to 1.}

\item{ypos_table}{The y position of the timeseries table with the basic statistics. Takes values from 0 to 1.}

\item{nbins}{The number of bins to split the data from the PDF calculation. Default is 30.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored when computing the statistics. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}

\item{plot}{A logical value, if TRUE the timeseries/PDF/ECDF/ACF panel is built. Default is FALSE.
For a multi-column input one panel is built per column.}
}
\value{
A list with two elements: \code{stats_table}, a data.frame of statistics
(rows are the statistics, columns are the input series - named after the input columns
when more than one is supplied, otherwise \code{"Value"}); and \code{plot}, which is
\code{NULL} when \code{plot = FALSE}, a single combined plot for a single-column input,
or a named list of one combined plot per column for a multi-column input.
}
\description{
This function provides basic statistics and plots of a timeseries in xts format.
It plots the timeseries values, the empirical probability density function (PDF),
the empirical cumulative density function (CDF) and the autocorrelation. Additionally, it will calculate
all major statistics.
\itemize{
\item Number of Data - Number of missing data - Percentage of missing Data
\item Minimum - Maximum - Average - Variance - Coefficient of Variation - Standard Deviation - Third moment - Skewness
\itemize{
\item Kurtosis
\item L-mean, L-scale , 3rd and 4th L-coefficients
\item Probability Dry
\item Quantiles 5th , 25th, 50th, 75th, 95th - Interquartile range.
\item Mean value and variance from a dry (defined by zero_threshold argument) to a wet state
\item Mean value and variance from a wet to a dry (defined by zero_threshold argument) state
\item Mean value and variance from a wet to a wet state
\item Transition probability from a wet to a wet state
\item Transition probability from a dry to a dry state
}
With arguments for plotting position of stats table and timeseries plotting period.
}
}
\examples{
file_path <- system.file("extdata", "KNMI_Daily.csv", package = "anyFit")
time_zone <- "UTC"
time_step <- "1 day"

data <- delim2xts(file_path = file_path,
                 time_zone = "UTC", delim = " ", time_step = time_step)

bstats <- basic_stats(data[,4], pstart = '2002',pend = '2005', show_table = FALSE,
show_label = FALSE, ignore_zeros = TRUE, plot = TRUE)

bstats$plot
bstats$stats_table

# Wide input: one column of statistics per series, plus one plot per series
wstats <- basic_stats(data[, 1:4], plot = TRUE)
wstats$stats_table          # columns named after the input series
wstats$plot[[1]]            # plot for the first series

}
