% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats_nc.R
\name{basic_stats_nc}
\alias{basic_stats_nc}
\title{basic_stats_nc}
\usage{
basic_stats_nc(
  raster_file,
  filename = NA,
  varname = NA,
  ignore_zeros = FALSE,
  zero_threshold = 0.01,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{raster_file}{A raster file}

\item{filename}{(optional) A NetCDF file name to import if raster_file is not provided.}

\item{varname}{(optional) The name of the variable to extract from 'filename'.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}

\item{parallel}{Logical, whether to use parallel processing.}

\item{...}{Additional arguments to pass to 'nc2xts' function (if 'filename' and 'varname' are provided).}
}
\value{
A raster object containing basic statistics.
}
\description{
This function calculates basic statistics (e.g., mean, min, max, SD) for a NetCDF raster file.
It returns the statistics in raster format.
}
