% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats_nc.R
\name{basic_stats_nc}
\alias{basic_stats_nc}
\title{basic_stats_nc}
\usage{
basic_stats_nc(
  raster_file,
  filename = NA,
  varname = NA,
  ignore_zeros = FALSE,
  zero_threshold = 0.01,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{raster_file}{A raster file}

\item{filename}{(optional) A NetCDF file name to import if raster_file is not provided.}

\item{varname}{(optional) The name of the variable to extract from 'filename'.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}

\item{parallel}{Logical, whether to use parallel processing.}

\item{...}{Additional arguments to pass to 'nc2xts' function (if 'filename' and 'varname' are provided).}
}
\value{
A raster object containing basic statistics.
}
\description{
This function calculates basic statistics (e.g., mean, min, max, SD) for a NetCDF raster file.
It returns the following statistics in raster format.
\itemize{
\item Number of Data - Number of missing data - Percentage of missing Data
\item Minimum - Maximum - Average - Variance - Coefficient of Variation - Standard Deviation - Third moment - Skewness
\itemize{
\item Kurtosis
\item L-mean, L-scale , 3rd and 4th L-coefficients
\item Probability Dry
\item Quantiles 5th , 25th, 50th, 75th, 95th - Interquartile range.
\item Mean value and variance from a dry (defined by zero_threshold argument) to a wet state
\item Mean value and variance from a wet to a dry (defined by zero_threshold argument) state
\item Mean value and variance from a wet to a wet state
\item Transition probability from a wet to a wet state
\item Transition probability from a dry to a dry state
}
}
}
\examples{
TO BE FILLED

}
