% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{BurXII}
\alias{BurXII}
\alias{dburr}
\alias{pburr}
\alias{qburr}
\alias{rburr}
\title{Burr Type XII Distribution}
\usage{
dburr(x, scale, shape1, shape2, PW = 1)

pburr(q, scale, shape1, shape2, PW = 1)

qburr(p, scale, shape1, shape2, PW = 1)

rburr(n, scale, shape1, shape2, PW = 1)
}
\description{
Burr Type XII distribution
}
