% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{check_missing}
\usage{
check_missing(data, periods, plot = TRUE, group_months = FALSE)
}
\arguments{
\item{data}{A xts object or matrix containing the timeseries data}

\item{periods}{A character vector containing one or more of the following:
"mins" or "minutes", "hours", "days", "weeks", "months", "quarters", and "years".}

\item{plot}{A logical value indicating whether to plot the missing values or not}

\item{group_months}{A logical value indicating whether to group the monthly missing values or not}
}
\value{
A list containing a data frame of the percentage of missing values for each period, and optionally a plot if set to TRUE
}
\description{
This function takes in time series data and one or more periods and checks for missing values.
It returns a data frame with the percentage of missing values for each period and a plot if set to TRUE.
The plot shows the percentage of missing values for each period.
If the period is set to "months" and "group_months" is set to TRUE,
the function returns also a plot of the percentage of missing values grouped by month.
}
\examples{
file_path <- system.file("extdata", "KNMI_Daily.csv", package = "anyFit")
time_zone <- "UTC"
time_step <- "1 day"

data <- delim2xts(file_path = file_path,
                 time_zone = "UTC", delim = " ", time_step = time_step)

missing <- check_missing(data[,seq(1,4)], c("months","years"))

missing$list_years$figure


}
