% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl_plots.R
\name{correl_plots}
\alias{correl_plots}
\title{correl_plots}
\usage{
correl_plots(
  x,
  y,
  check_common = TRUE,
  ignore_zeros = FALSE,
  zero_threshold = 0.01
)
}
\arguments{
\item{x, y}{A xts objects containing the time series data.}

\item{check_common}{logical value, if TRUE the function will test for common dates and excludes the rest.
Default is TRUE.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}
}
\value{
A list with the combined plots and the individual plots for greater flexibility.
}
\description{
Function that calculates correlation plots for two timeseries.
Scatter plot with Pearson correlation coefficient
Empirical copula plot with Spearman correlation coefficient
Standardized or Normal plot with Pearson semi-correlations
Handles missing values (NA) with casewise deletion
Can check and retain only common dates. Useful for cross-correlation between different stations and variables.
}
\examples{
file_path <- system.file("extdata", "KNMI_Daily.csv", package = "anyFit")
time_zone <- "UTC"
time_step <- "1 day"

data <- delim2xts(file_path = file_path,
                 time_zone = "UTC", delim = " ", time_step = time_step)

correls <- correl_plots(data[,1],data[,4], check_common = TRUE, ignore_zeros = TRUE)

correls$combined


}
