% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{Dagum}
\alias{Dagum}
\alias{pdagum}
\alias{ddagum}
\alias{qdagum}
\alias{rdagum}
\title{Dagum Distribution}
\usage{
pdagum(q, scale, shape1, shape2, PW = 1)

ddagum(q, scale, shape1, shape2, PW = 1)

qdagum(p, scale, shape1, shape2, PW = 1)

rdagum(n, scale, shape1, shape2, PW = 1)
}
\description{
Dagum distribution
}
