% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delim2xts.R
\name{delim2xts}
\alias{delim2xts}
\title{delim2xts}
\usage{
delim2xts(
  file_path,
  time_zone,
  date_index = 1,
  strict_step = TRUE,
  delim = "\\t",
  skip_rows = 0,
  col_names = TRUE,
  no_value = NA,
  exc_leaps = TRUE,
  save_Xts = FALSE,
  filename = NA
)
}
\arguments{
\item{file_path}{Path to the file with the data.}

\item{time_zone}{Time zone of the data.}

\item{date_index}{Position of the date column in the data file.}

\item{strict_step}{Logical that indicates if the timeseries are recorded in strict timestep.}

\item{delim}{Delimiter used in the data file.}

\item{skip_rows}{Number of rows to be skipped in the data file.}

\item{col_names}{Logical that indicates if the data file has column names.}

\item{no_value}{Value used as NA.}

\item{exc_leaps}{Logical that indicates if leap years shall be excluded.}

\item{save_Xts}{Logical that indicates if the xts shall be saved as a txt file.}

\item{filename}{Name of the file where the xts will be saved.}
}
\value{
An xts object.
}
\description{
A function to convert data from delimited files to xts objects.
The function takes a file path, time zone, date index, time step, delimiter,
skip rows, column names, no value, and exc_leaps (exclude 29th of February from leap years) as arguments.
If the time step of the timeseries is strict then the function will try to identify potential gaps in the data and indicate them with NANs.
Otherwise the function will return the raw dataset.
The user is given the option to save the xts dataset in a text file.
}
\examples{
file_path <- system.file("extdata", "KNMI_Daily.csv", package = "anyFit")
time_zone <- "UTC"

data <- delim2xts(file_path = file_path,
                 time_zone = "UTC", delim = " ", strict_step = TRUE)

}
