% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{Exponential}
\alias{Exponential}
\alias{dexp}
\alias{pexp}
\alias{qexp}
\alias{rexp}
\title{Exponential Distribution}
\usage{
dexp(x, location, scale)

pexp(q, location, scale)

qexp(p, location, scale)

rexp(n, location, scale)
}
\description{
Exponential distribution
}
