% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{ExpWeibull}
\alias{ExpWeibull}
\alias{pexpweibull}
\alias{dexpweibull}
\alias{qexpweibull}
\alias{rexpweibull}
\title{Exponential Weibull Distribution}
\usage{
pexpweibull(q, scale, shape1, shape2, log.p = FALSE)

dexpweibull(x, scale, shape1, shape2, log = FALSE)

qexpweibull(p, scale, shape1, shape2)

rexpweibull(n, scale, shape1, shape2)
}
\description{
Exponential Weibull distribution
}
