% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ACF.R
\name{fit_ACF}
\alias{fit_ACF}
\title{fit_ACF}
\usage{
fit_ACF(
  ts,
  lag_max,
  type = list("CAS", "HK", "SRD"),
  ignore_zeros = FALSE,
  zero_threshold = 0.01
)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{lag_max}{Maximum lag to use in fitting.}

\item{type}{A list of character strings, containing the type of ACF to fit.
Options include: \code{'CAS'}, \code{'HK'}, and \code{'SRD'}.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}
}
\value{
A list containing a vector of fitted parameters, a data frame of
fitted ACF values, and a plot of the fitted ACFs.
}
\description{
Function to fit theoretical Autocorrelation Functions (ACFs) to timeseries data.
Three functions are available, the Cauchy-type autocorrelation structure,
the Hurst - Kolmogorov and the Short Range dependence structure.
}
