% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_diagnostics.R
\name{fit_diagnostics}
\alias{fit_diagnostics}
\title{fit_diagnostics}
\usage{
fit_diagnostics(
  ts,
  dist = "norm",
  params,
  ignore_zeros = FALSE,
  zero_threshold = 0.01
)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{params}{A list object containing the parameters of the fitted distribution.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}

\item{distr}{The distribution function to be tested.}
}
\value{
A list containing a Q-Q plot and a P-P plot and a goodness-of-fit table.
The GoF metric calculated are the Kramer von-Mises and Kolmogorov-Smirnov. NEEDS EXPANSION
}
\description{
Performs a goodness-of-fit test to diagnose fitted distributions.
}
