% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{fitlm_dagum}
\alias{fitlm_dagum}
\title{fitlm_dagum}
\usage{
fitlm_dagum(x, ignore_zeros = FALSE, zero_threshold = 0.01, order = c(1:5))
}
\arguments{
\item{x}{A xts object containing the time series data.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}

\item{method}{The method to determine the starting values for the fitting. Can be 'knn' or 'DEoptim'. Default is 'knn'. See description for details.}
}
\description{
Function for fitting the Dagum distribution using the L-Moments method.
Since there is not closed expression for the L-Moments of this distribution, the fitting must be done numerically.
For this purpose we employ the lmom::pelp function which works by optimization and requires an initial set of parameters.
}
