% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{fitlm_gengamma}
\alias{fitlm_gengamma}
\title{fitlm_gengamma}
\usage{
fitlm_gengamma(x, ignore_zeros = FALSE, zero_threshold = 0.01)
}
\arguments{
\item{x}{A xts object containing the time series data.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}
}
\description{
Function for fitting the Generalized Gamma distribution using the L-Moments method
}
