% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{fitlm_GPD}
\alias{fitlm_GPD}
\title{fitlm_GPD}
\usage{
fitlm_GPD(x, bound = NULL, ignore_zeros = FALSE, zero_threshold = 0.01)
}
\arguments{
\item{x}{A xts object containing the time series data.}

\item{bound}{Is the distribution bound? Default is NULL.}
}
\description{
Function for fitting the Generalized Pareto Distribution using the L-Moments method
}
