% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlm_nc.R
\name{fitlm_nc}
\alias{fitlm_nc}
\title{fitlm_nc}
\usage{
fitlm_nc(
  raster_file,
  filename = NA,
  varname = NA,
  candidates = "norm",
  ignore_zeros = FALSE,
  zero_threshold = 0.01,
  parallel = FALSE,
  ncores = 2,
  ...
)
}
\arguments{
\item{raster_file}{A raster file.}

\item{filename}{(optional) A NetCDF file name to import if raster_file is not provided.}

\item{varname}{(optional) The name of the variable to extract from 'filename'.}

\item{candidates}{A list of distribution to fit.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}

\item{parallel}{Logical, whether to use parallel processing.}

\item{ncores}{Number of cores to use in the case of parallel computations}

\item{...}{Additional arguments to pass to 'nc2xts' function (if 'filename' and 'varname' are provided).}
}
\value{
A list of raster objects containing the fitted distribution parameters, the theoretical L-moments of the fitted distribution,
the sample L-Moments and goodness-of-fit statistics in raster format.
}
\description{
This function fits a list of candidate distributions using the L-moments method NetCDF raster file.
It returns fitted distribution parameters, the theoretical L-moments of the fitted distribution,
the sample L-Moments and goodness-of-fit statistics in raster format.
}
\examples{
TO BE FILLED

}
