% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlm_nxts.R
\name{fitlm_nxts}
\alias{fitlm_nxts}
\title{fitlm_nxts}
\usage{
fitlm_nxts(
  ts,
  candidates,
  nrow = 5,
  ncol = 4,
  ignore_zeros = FALSE,
  zero_threshold = 0.01
)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{candidates}{A list of distribution to fit.}

\item{nrow}{Number of rows for plotting.}

\item{ncol}{Number of columns for plotting.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}
}
\value{
A list with the estimated parameters, diagnostic plots, QQ plots and PP plots.
}
\description{
This function fits a list of candidate distributions using the L-moments method
to an arbitrary number of timeseries in xts format. Additionally to the list of the fitted parameters,
goodness-of-fit metric, PP and QQ plots.
}
\examples{
file <- "KNMI_Daily.csv"
file_path <- file.path(cwd,file)
time_zone <- "UTC"
time_step <- "1 day"

data <- delim2xts(file_path = file_path, 
                 time_zone = "UTC", delim = " ", time_step = time_step)
                
candidates <- list('exp','expweibull', 'gamma3')
                                     
fits_all <- fitlm_nxts(data, candidates, nrow = 5, ncol = 4, ignore_zeros = TRUE)
fits_all$QQ_panels[[1]]                 

}
