% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{Gamma}
\alias{Gamma}
\alias{dgamma}
\alias{pgamma}
\alias{qgamma}
\alias{rgamma}
\title{Gamma Distribution}
\usage{
dgamma(x, scale, shape)

pgamma(q, scale, shape)

qgamma(p, scale, shape)

rgamma(n, scale, shape)
}
\description{
Gamma distribution
}
