% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{Gamma3}
\alias{Gamma3}
\alias{dgamma3}
\alias{pgamma3}
\alias{qgamma3}
\alias{rgamma3}
\title{3-parameter Gamma Distribution}
\usage{
dgamma3(x, location, scale, shape)

pgamma3(q, location, scale, shape)

qgamma3(p, location, scale, shape)

rgamma3(n, location, scale, shape)
}
\description{
3-parameter Gamma distribution
}
