% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{GenGamma-Location}
\alias{GenGamma-Location}
\alias{pgengamma_loc}
\alias{dgengamma_loc}
\alias{qgengamma_loc}
\alias{rgengamma_loc}
\title{Generalized Gamma with Location Distribution}
\usage{
pgengamma_loc(q, location, scale, shape1, shape2)

dgengamma_loc(x, location, scale, shape1, shape2)

qgengamma_loc(p, location, scale, shape1, shape2)

rgengamma_loc(n, location, scale, shape1, shape2)
}
\description{
Generalized Gamma with Location distribution
}
