% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{GenGamma}
\alias{GenGamma}
\alias{dgengamma}
\alias{pgengamma}
\alias{qgengamma}
\alias{rgengamma}
\title{Generalized Gamma Distribution}
\usage{
dgengamma(x, scale, shape1, shape2)

pgengamma(q, scale, shape1, shape2)

qgengamma(p, scale, shape1, shape2)

rgengamma(n, scale, shape1, shape2)
}
\description{
Generalized Gamma distribution
}
