% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{GEV}
\alias{GEV}
\alias{dgev}
\alias{pgev}
\alias{qgev}
\alias{rgev}
\title{Generalized Extreme Value Distribution}
\usage{
dgev(x, location, scale, shape)

pgev(q, location, scale, shape)

qgev(p, shape, scale, location)

rgev(n, location, scale, shape)
}
\description{
Generalized Extreme Value distribution
}
