% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{Gumbel}
\alias{Gumbel}
\alias{dgumbel}
\alias{pgumbel}
\alias{qgumbel}
\alias{rgumbel}
\title{Gumbel Distribution}
\usage{
dgumbel(x, location, scale)

pgumbel(q, location, scale)

qgumbel(p, location, scale)

rgumbel(n, location, scale)
}
\description{
Gumbel distribution
}
