% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmom_stats.R
\name{lmom_stats}
\alias{lmom_stats}
\title{lmom_stats}
\usage{
lmom_stats(ts, ignore_zeros = FALSE, zero_threshold = 0.01)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}
}
\value{
A dataframe with the L-Mean, L-Scale, L-Skew, L-Kurtosis and L-CV of the time series
}
\description{
Calculate the L-Mean, L-Scale, L-Skew, L-Kurtosis and L-CV of a time series.
}
\examples{
file_path <- system.file("extdata", "KNMI_Daily.csv", package = "anyFit")
time_zone <- "UTC"
time_step <- "1 day"

data <- delim2xts(file_path = file_path,
                 time_zone = "UTC", delim = " ", time_step = time_step)
ts_lmoms <- lmom_stats(data, ignore_zeros = TRUE)


}
