% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{Log-Normal}
\alias{Log-Normal}
\alias{dlognorm}
\alias{plognorm}
\alias{qlognorm}
\alias{rlognorm}
\title{Log-Normal Distribution}
\usage{
dlognorm(x, location = 0, scale, shape)

plognorm(q, location = 0, scale, shape)

qlognorm(p, location = 0, scale, shape)

rlognorm(n, location = 0, scale, shape)
}
\description{
Log-Normal distribution
}
