% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRatio_check.R
\name{LRatio_check}
\alias{LRatio_check}
\title{LRatio_check}
\usage{
LRatio_check(ts_lmoms)
}
\arguments{
\item{ts_lmoms}{A vector or a matrix of L-moments of the time series of interest.
Can test multiple timeseries as columns of a matrix.}
}
\value{
A list containing a boolean (TRUE/FALSE) vector with the results of the test
and a ggplot object with the visual representation of the test.
}
\description{
This function checks if the L-ratios of a given time series
are within the acceptable parameter space for four different distributions.
These are the Dagum, GGamma, ExpWeibull, and BurrXII. TO BE EXANDED
}
\examples{

file_path <- system.file("extdata", "KNMI_Daily.csv", package = "anyFit")
time_zone <- "UTC"
time_step <- "1 day"

data <- delim2xts(file_path = file_path,
                 time_zone = "UTC", delim = " ", time_step = time_step)
ts_lmoms <- lmom_stats(data, ignore_zeros = TRUE)

lcheck <- LRatio_check(ts_lmoms)

lcheck$multi_plots

}
