% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_boxplots.R
\name{monthly_boxplots}
\alias{monthly_boxplots}
\title{monthly_boxplots}
\usage{
monthly_boxplots(
  ts,
  palette = "Set1",
  ignore_zeros = FALSE,
  zero_threshold = 0.01
)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{palette}{A color palette. Default is 'Set1'.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}
}
\value{
A ggplot object with monthly boxplots
}
\description{
Plot monthly boxplots for a given time series.
}
