% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_ecdf.R
\name{monthly_ecdf}
\alias{monthly_ecdf}
\title{monthly_ecdf}
\usage{
monthly_ecdf(ts, ignore_zeros = FALSE, zero_threshold = 0.01)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored when computing the statistics. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}
}
\value{
A grid of monthly ECDF plots.
}
\description{
This function takes a time series and generates a grid of monthly Empirical
Cumulative Distribution Function (ECDF) plots for each month in the time series.
}
