% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_fdiagnostics.R
\name{monthly_fdiagnostics}
\alias{monthly_fdiagnostics}
\title{monthly_fdiagnostics}
\usage{
monthly_fdiagnostics(
  ts,
  distr,
  params,
  ignore_zeros = FALSE,
  zero_threshold = 0.01,
  nrow = 3,
  ncol = 4
)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{distr}{The distribution function to be tested.}

\item{params}{A list object containing the parameters of the fitted distribution.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}

\item{nrow}{Number of rows of the plots.}

\item{ncol}{Number of columns of the plots.}
}
\value{
A list containing a Q-Q plot and a P-P plot for each month and a goodness-of-fit table.
The GoF metric calculated are the Kramer von-Mises and Kolmogorov-Smirnov. NEEDS EXPANSION
}
\description{
Performs a goodness-of-fit test to diagnose fitted distributions by month.
}
