% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_stats.R
\name{monthly_stats}
\alias{monthly_stats}
\title{monthly_stats}
\usage{
monthly_stats(
  ts,
  aggregated = FALSE,
  FUN = "mean",
  ignore_zeros = FALSE,
  zero_threshold = 0.01,
  title = FALSE
)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{aggregated}{A logical value, if TRUE the statistics are also given on the aggregated monthly scale
of the time series. Default is FALSE.}

\item{FUN}{The aggregation function applied to the time series. Default is 'mean'.}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored when computing the statistics. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}

\item{title}{A logical value, if TRUE then a title will be added to the plots. Default is FALSE.}
}
\value{
If aggregated is TRUE, a list containing the aggregated statistics and the plots is returned.
If aggregated is FALSE, a list containing the base scale statistics and the plots is returned.
}
\description{
This function provides monthly statistics in aggregated and original (base) timescale
and plots of a time series. Statistics ploted are average, standard deviation, skewness. For the base scale
the probability dry is calculated. For the monthly scale the month to month correlations are calculated.
Month to month correlations are calculated backwards. i.e January == cor(December, January). Statistics calculated
and provided on a table are the number of data points,
number of missing data, percentage of missing data, min, max, mean, variance,
standard deviation, variation, 3rd moment, skewness, kurtosis and probability dry.
}
\examples{
file_path <- system.file("extdata", "KNMI_Daily.csv", package = "anyFit")
time_zone <- "UTC"
time_step <- "1 day"

data <- delim2xts(file_path = file_path,
                 time_zone = "UTC", delim = " ", time_step = time_step)

monthly_sts <- monthly_stats(data[,4],base_scale = TRUE, FUN = "sum")

monthly_sts$fbase
monthly_sts$base_list$July$raw_stats

}
