% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc2xts.R
\name{nc2xts}
\alias{nc2xts}
\title{nc2xts

This function reads a NetCDF raster file, optionally masks it by country, continent, or user defined shapefile,
and converts it into an xts object for further analysis.}
\usage{
nc2xts(filename, varname, shapefile = NA, country = NA, continent = NA)
}
\arguments{
\item{filename}{The path to the NetCDF file.}

\item{varname}{The name of the variable to extract from the NetCDF file.}

\item{shapefile}{(optional) The path to a shapefile for masking.}

\item{country}{(optional) The name of a country for masking.}

\item{continent}{(optional) The name of a continent for masking.}
}
\value{
A list containing the raster, xts object, dates, and coordinates.
}
\description{
nc2xts

This function reads a NetCDF raster file, optionally masks it by country, continent, or user defined shapefile,
and converts it into an xts object for further analysis.
}
\examples{
# Load required libraries
library(ncdf4)
library(raster)
library(xts)

# Example usage 1: Convert NetCDF raster to xts object
nc_file <-"data/ncfile.nc"
result <- nc2xts(filename = nc_file, varname = "tp")
nc_ggplot(result$raster)

# Example usage 2: Spatial subsetting by country
nc_file <-"data/ncfile.nc"
result <- nc2xts(filename = nc_file, varname = "tp", country = "France")
nc_ggplot(result$raster)

}
