% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc2xts_nn.R
\name{nc2xts_nn}
\alias{nc2xts_nn}
\title{nc2xts}
\usage{
nc2xts_nn(filename, varname, coords)
}
\arguments{
\item{filename}{The path to the NetCDF file. It can be a character vector of multiple NetCDF files.
This is useful when data are downloaded in time chunks. The function will automatically order by date.}

\item{varname}{A character vector of variable names to extract.}

\item{coords}{A matrix or data.frame of coordinates (longitude and latitude) for extraction.}
}
\value{
An xts time series object containing the extracted variables.
}
\description{
This function extracts one or more variables from NetCDF files from one or more locations at specified coordinates
and creates an xts time series. This function performs no interpolation/extrapolation of the data.
It return the values of the cell which contains the point coordinates. If the specified points fall outside of the dataset
it will return NANs.
}
\examples{

rnd_xts = nc2xts_nn(filename = filename, varname = "rr", coords = coords)

}
