% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_ggplot.R
\name{nc_ggplot}
\alias{nc_ggplot}
\title{nc_ggplot}
\usage{
nc_ggplot(raster_file)
}
\arguments{
\item{raster_file}{A NetCDF raster file.}
}
\value{
A ggplot2 object representing the raster plot.
}
\description{
This function takes a NetCDF raster file and a variable name as inputs, and generates
a ggplot2 raster plot.
}
\examples{

# Example usage:
# Create a ggplot2 raster plot from a NetCDF raster file
nc_file <- system.file("extdata/ncfile.nc", package = "raster")
plot <- nc_ggplot(nc_file, "temperature")
plot

}
