% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_ggplot.R
\name{nc_ggplot}
\alias{nc_ggplot}
\title{nc_ggplot}
\usage{
nc_ggplot(
  raster_file,
  title = FALSE,
  legend.title = NA,
  viridis.option = "viridis",
  ...
)
}
\arguments{
\item{raster_file}{A raster file.}

\item{title}{Logical, whether to add variable names as plot titles.}

\item{legend.title}{The title for the color legend.}

\item{viridis.option}{The viridis color palette option (e.g., "viridis", "magma", "plasma").}

\item{...}{Additional arguments to pass to 'wrap_plots' function from the 'patchwork' package.}
}
\value{
A ggplot2 object representing the raster plot.
}
\description{
This function takes a NetCDF raster file and generates
a ggplot2 raster plot. Supports RasterBricks e.g. Timeseries of rasters
}
\examples{

# Example usage:
# Create a ggplot2 raster plot from a NetCDF raster file
nc_file <- system.file("extdata/ncfile.nc", package = "raster")
plot <- nc_ggplot(nc_file, "temperature")
plot

}
