% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{Normal}
\alias{Normal}
\alias{dnorm}
\alias{pnorm}
\alias{qnorm}
\alias{rnorm}
\title{Normal Distribution}
\usage{
dnorm(x, mean = 0, sd = 1)

pnorm(q, mean = 0, sd = 1)

qnorm(p, mean = 0, sd = 1)

rnorm(n, mean = 0, sd = 1)
}
\description{
Normal distribution
}
