% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_xts.R
\name{normalise_xts}
\alias{normalise_xts}
\title{normalise_xts}
\usage{
normalise_xts(
  ts,
  dist_period = "monthly",
  ignore_zeros = FALSE,
  zero_threshold = 0.01
)
}
\arguments{
\item{ts}{A xts object containing the time series data.}

\item{dist_period}{The distribution period for normalization ('monthly' or NA).}

\item{ignore_zeros}{A logical value, if TRUE zeros will be ignored when computing the statistics. Default is FALSE.}

\item{zero_threshold}{The threshold below which values are considered zero. Default is 0.01.}
}
\value{
A list of normalized xts time series.
}
\description{
This function normalizes an xts time series by transforming the data into standard normal
distribution, either on a monthly basis or for the entire time series.
Tranforming the data in the standard normal space can assist in visually identifying outliers.
Using the monthly distribution to normalise the data can be useful in cyclo-stationary processes.
In other cases, 'dist_period' should be set to NA.
}
\examples{
TO BE FILLED

}
