% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_apply_nc.R
\name{period_apply_nc}
\alias{period_apply_nc}
\title{period_apply_nc

This function applies a summary function (e.g., mean, sum) to a NetCDF raster file
over specified time periods (e.g., months, years).}
\usage{
period_apply_nc(
  raster_file,
  filename = NA,
  varname = NA,
  period = "months",
  period_multiplier = 1,
  FUN = "mean",
  ...
)
}
\arguments{
\item{raster_file}{A raster file}

\item{filename}{(optional) A NetCDF file name to import if raster_file is not provided.}

\item{varname}{(optional) The name of the variable to extract from 'filename'.}

\item{period}{The time period to apply the function (e.g., "months", "years").}

\item{period_multiplier}{a multiplier to define arbitrary periods based on the period argument}

\item{FUN}{The summary function to apply (e.g., "mean", "sum").}

\item{...}{Additional arguments to pass to 'nc2xts' function (if 'filename' and 'varname' are provided).}
}
\value{
A raster object with the summary statistics for the specified time periods.
}
\description{
period_apply_nc

This function applies a summary function (e.g., mean, sum) to a NetCDF raster file
over specified time periods (e.g., months, years).
}
\examples{
# Load required libraries
library(raster)
library(xts)

# Example usage 1: Using a raster file directly
result <- period_apply_nc(raster, period = "months", FUN = "mean")
nc_ggplot(result)

# Example usage 2: Using a filename and variable name
nc_file <-"data/ncfile.nc"
result <- period_apply_nc(filename = nc_file, varname = "tp", period = "months", FUN = "sum")
nc_ggplot(result)

}
