% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{plot_missing}
\usage{
plot_missing(ts)
}
\arguments{
\item{ts}{An xts time series object.}
}
\value{
A \code{ggplot} object with the missing values.
}
\description{
Function to plot missing values position.
Assigns the mean value of timeseries to NA for easy identification of empties.
Meant for quick and dirty (i.e. Engineering) visual checks
}
\examples{
file_path <- system.file("extdata", "KNMI_Daily.csv", package = "anyFit")
time_zone <- "UTC"
time_step <- "1 day"

data <- delim2xts(file_path = file_path,
                 time_zone = "UTC", delim = " ", time_step = time_step)


plot_missing(data[,seq(1,4)])

}
