% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{Rayleigh}
\alias{Rayleigh}
\alias{drayleigh}
\alias{prayleigh}
\alias{qrayleigh}
\alias{rrayleigh}
\title{Rayleigh Distribution}
\usage{
drayleigh(x, location, scale)

prayleigh(x, location, scale)

qrayleigh(p, location, scale)

rrayleigh(n, location, scale)
}
\description{
Rayleigh distribution
}
