% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_xts.R
\name{ridge_plots}
\alias{ridge_plots}
\title{ridge_xts}
\usage{
ridge_plots(
  data,
  palette = "Set3",
  ignore_zeros = FALSE,
  zero_threshold = 0.01
)
}
\arguments{
\item{data}{A vector or matrix of data}

\item{palette}{Color palette to use for ridge plot. Default is 'Set3'}

\item{ignore_zeros}{Boolean specifying whether to ignore values below zero_threshold}

\item{zero_threshold}{Threshold value for ignoring data. Default is 0.01}
}
\value{
A list with two elements: plot_all and plot_monthly. plot_all is the full ridge plot for the data, plot_monthly is a list of ridge plots for each month of the data.
}
\description{
ridge_xts
}
\examples{
data(airquality)
ridge_plots(airquality$Ozone)

}
