% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxts.R
\name{sxts}
\alias{sxts}
\alias{str.sxts}
\alias{summary.sxts}
\alias{print.sxts}
\alias{attributes.sxts}
\alias{coords}
\alias{coords.sxts}
\alias{projection}
\alias{projection.sxts}
\alias{elements}
\alias{elements.sxts}
\alias{[.sxts}
\alias{is.sxts}
\alias{lag.sxts}
\alias{diff.sxts}
\alias{Ops.sxts}
\alias{mask.sxts}
\alias{rasterFromSxts.sxts}
\alias{sxtsFromRaster}
\alias{rbindlist.sxts}
\title{sxts Class - Spatial eXtensible Time Series}
\usage{
sxts(data, order.by, coords = NULL, projection = NULL)

\method{str}{sxts}(object, ...)

\method{summary}{sxts}(obj, ...)

\method{print}{sxts}(x, ...)

attributes.sxts(obj)

coords(x)

\method{coords}{sxts}(x)

projection(x)

\method{projection}{sxts}(x)

elements(x)

\method{elements}{sxts}(x)

\method{[}{sxts}(x, i, j, drop = FALSE, ...)

is.sxts(x)

\method{lag}{sxts}(x, k = 1, ...)

\method{diff}{sxts}(x, lag = 1, differences = 1, ...)

\method{Ops}{sxts}(e1, e2)

mask.sxts(obj, xlim = NA, ylim = NA, shapefile_name = NA, mask = NA)

\method{rasterFromSxts}{sxts}(obj, ...)

sxtsFromRaster(raster, ...)

rbindlist.sxts(sxts_list)
}
\arguments{
\item{data}{Matrix or data.frame of time series data}

\item{order.by}{POSIXct vector for time index}

\item{coords}{A data.frame or matrix with coordinates (x, y columns)}

\item{projection}{Character string specifying the coordinate reference system}

\item{...}{Additional arguments passed to xts()}

\item{sxts_list}{A list of \code{sxts} objects sharing the same columns (spatial
locations). Coordinates and projection are taken from the first element.}
}
\value{
An sxts object
}
\description{
A spatial extension of xts class using attributes to store
spatial information. This approach is more elegant and memory-efficient.
Create a spatial xts (sxts) object

Drop-in replacement for \code{do.call(rbind.xts, list)} that is much faster and
lighter on memory. \code{rbind.xts} merges objects pairwise, re-checking the index
and copying attributes at every step. This collects the bare coordinate data
once and rebuilds a single \code{sxts} object, so the time index is concatenated
and the data matrices are bound in one pass.
}
