% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxts.R
\name{sxts}
\alias{sxts}
\alias{str.sxts}
\alias{summary.sxts}
\alias{print.sxts}
\alias{attributes.sxts}
\alias{coords}
\alias{coords.sxts}
\alias{projection}
\alias{projection.sxts}
\alias{elements}
\alias{elements.sxts}
\alias{[.sxts}
\alias{is.sxts}
\title{sxts Class - Spatial eXtensible Time Series}
\usage{
sxts(data, order.by, coords = NULL, projection = NULL)

\method{str}{sxts}(object, ...)

summary.sxts(obj, ...)

print.sxts(x, ...)

attributes.sxts(obj)

coords(x)

\method{coords}{sxts}(x)

projection(x)

\method{projection}{sxts}(x)

elements(x)

\method{elements}{sxts}(x)

\method{[}{sxts}(x, i, j, drop = FALSE, ...)

is.sxts(x)
}
\arguments{
\item{data}{Matrix or data.frame of time series data}

\item{order.by}{POSIXct vector for time index}

\item{coords}{A data.frame or matrix with coordinates (x, y columns)}

\item{projection}{Character string specifying the coordinate reference system}

\item{...}{Additional arguments passed to xts()}
}
\value{
An sxts object
}
\description{
A spatial extension of xts class using attributes to store
spatial information. This approach is more elegant and memory-efficient.
Create a spatial xts (sxts) object
}
