% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_Anyfit.R
\name{Weibull3}
\alias{Weibull3}
\alias{dweibull}
\alias{pweibull}
\alias{qweibull}
\alias{rweibull}
\title{3-parameter Weibull Distribution}
\usage{
dweibull(x, location, scale, shape)

pweibull(q, location, scale, shape)

qweibull(p, location, scale, shape)

rweibull(n, location, scale, shape)
}
\description{
3-parameter Weibull distribution
}
